add base_det3d_inferencer
