+++ mmdet3d/apis/__init__.py
@@ -3,10 +3,10 @@ from .inference import (convert_SyncBN, inference_detector,
                         inference_mono_3d_detector,
                         inference_multi_modality_detector, inference_segmentor,
                         init_model)
-from .inferencers import Mono3DDetInferencer
+from .inferencers import BaseDet3DInferencer, MonoDet3DInferencer
 
 __all__ = [
     'inference_detector', 'init_model', 'inference_mono_3d_detector',
     'convert_SyncBN', 'inference_multi_modality_detector',
-    'inference_segmentor', 'Mono3DDetInferencer'
+    'inference_segmentor', 'BaseDet3DInferencer', 'MonoDet3DInferencer'
 ]
