+++ mmdet3d/apis/inferencers/base_det3d_inferencer.py
@@ -1,12 +1,9 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-import os.path as osp
 from typing import Dict, List, Optional, Sequence, Tuple, Union
 
-import mmcv
 import mmengine
 import numpy as np
 import torch.nn as nn
-from mmengine.dataset import Compose
 from mmengine.infer.infer import BaseInferencer, ModelType
 from mmengine.runner import load_checkpoint
 from mmengine.structures import InstanceData
@@ -39,8 +36,8 @@ def convert_SyncBN(config):
                 convert_SyncBN(config[item])
 
 
-class Mono3DDetInferencer(BaseInferencer):
-    """MMDet3D Mono3D inferencer.
+class BaseDet3DInferencer(BaseInferencer):
+    """Base 3D object detection inferencer.
 
     Args:
         model (str, optional): Path to the config file or the model name
@@ -70,9 +67,6 @@ class Mono3DDetInferencer(BaseInferencer):
                  device: Optional[str] = None,
                  scope: Optional[str] = 'mmdet3d',
                  palette: str = 'none') -> None:
-        # A global counter tracking the number of images processed, for
-        # naming of the output images
-        self.num_visualized_imgs = 0
         self.palette = palette
         register_all_modules()
         super().__init__(
@@ -112,18 +106,6 @@ class Mono3DDetInferencer(BaseInferencer):
         model.to(device)
         model.eval()
         return model
-
-    def _init_pipeline(self, cfg: ConfigType) -> Compose:
-        """Initialize the test pipeline."""
-        pipeline_cfg = cfg.test_dataloader.dataset.pipeline
-
-        load_img_idx = self._get_transform_idx(pipeline_cfg,
-                                               'LoadImageFromFileMono3D')
-        if load_img_idx == -1:
-            raise ValueError(
-                'LoadImageFromFileMono3D is not found in the test pipeline')
-        pipeline_cfg[load_img_idx]['type'] = 'Mono3DInferencerLoader'
-        return Compose(pipeline_cfg)
 
     def _get_transform_idx(self, pipeline_cfg: ConfigType, name: str) -> int:
         """Returns the index of the transform in a pipeline.
@@ -194,80 +176,6 @@ class Mono3DDetInferencer(BaseInferencer):
             print_result=print_result,
             pred_out_file=pred_out_file,
             **kwargs)
-
-    def visualize(self,
-                  inputs: InputsType,
-                  preds: PredType,
-                  return_vis: bool = False,
-                  show: bool = False,
-                  wait_time: int = 0,
-                  draw_pred: bool = True,
-                  pred_score_thr: float = 0.3,
-                  img_out_dir: str = '') -> Union[List[np.ndarray], None]:
-        """Visualize predictions.
-
-        Args:
-            inputs (List[Union[str, np.ndarray]]): Inputs for the inferencer.
-            preds (List[Dict]): Predictions of the model.
-            return_vis (bool): Whether to return the visualization result.
-                Defaults to False.
-            show (bool): Whether to display the image in a popup window.
-                Defaults to False.
-            wait_time (float): The interval of show (s). Defaults to 0.
-            draw_pred (bool): Whether to draw predicted bounding boxes.
-                Defaults to True.
-            pred_score_thr (float): Minimum score of bboxes to draw.
-                Defaults to 0.3.
-            img_out_dir (str): Output directory of visualization results.
-                If left as empty, no file will be saved. Defaults to ''.
-        Returns:
-            List[np.ndarray] or None: Returns visualization results only if
-            applicable.
-        """
-        if self.visualizer is None or (not show and img_out_dir == ''
-                                       and not return_vis):
-            return None
-
-        if getattr(self, 'visualizer') is None:
-            raise ValueError('Visualization needs the "visualizer" term'
-                             'defined in the config, but got None.')
-
-        results = []
-
-        for single_input, pred in zip(inputs, preds):
-            if isinstance(single_input, str):
-                img_bytes = mmengine.fileio.get(single_input)
-                img = mmcv.imfrombytes(img_bytes)
-                img = img[:, :, ::-1]
-                img_name = osp.basename(single_input)
-            elif isinstance(single_input, np.ndarray):
-                img = single_input.copy()
-                img_num = str(self.num_visualized_imgs).zfill(8)
-                img_name = f'{img_num}.jpg'
-            else:
-                raise ValueError('Unsupported input type: '
-                                 f'{type(single_input)}')
-
-            out_file = osp.join(img_out_dir, img_name) if img_out_dir != '' \
-                else None
-
-            data_input = dict(img=img)
-            self.visualizer.add_datasample(
-                img_name,
-                data_input,
-                pred,
-                show=show,
-                wait_time=wait_time,
-                draw_gt=False,
-                draw_pred=draw_pred,
-                pred_score_thr=pred_score_thr,
-                out_file=out_file,
-                vis_task='mono_det',
-            )
-            results.append(img)
-            self.num_visualized_imgs += 1
-
-        return results
 
     def postprocess(
         self,
